requirements.txt
torch
torchvision
kaggle
matplotlib
pandas
numpy
gradio
scikit-learn
Pillow
seaborn
